requirements.txt
asgiref==3.2.10
coverage==5.2.1
Django==3.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
python-stdnum==1.14
pytz==2020.1
rut-chile==1.2.0
sqlparse==0.3.1
